% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFBS-DYNBPS-v2.R
\name{make_prior}
\alias{make_prior}
\title{Convert a full prior covariance matrix to block form for spFFBS}
\usage{
make_prior(m0, C0, nu0, Psi0, p)
}
\arguments{
\item{m0}{(p+n) x q prior mean matrix.}

\item{C0}{(p+n) x (p+n) prior covariance matrix (full).}

\item{nu0}{Prior degrees of freedom (scalar).}

\item{Psi0}{q x q prior scale matrix.}

\item{p}{Number of regression coefficients (integer).}
}
\value{
A list with elements m, C_bb, C_bw, C_ww, nu, Psi
ready to pass as \code{prior} to \code{spFFBS}.
}
\description{
Convert a full prior covariance matrix to block form for spFFBS
}
