% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spFFBS-package.R
\docType{package}
\name{spFFBS-package}
\alias{spFFBS-package}
\title{spFFBS: Spatiotemporal Propagation for Multivariate Bayesian Dynamic Learning}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implementation of the Forward Filtering Backward Sampling (FFBS) algorithm with Dynamic Bayesian Predictive Stacking (DYNBPS) integration for multivariate spatiotemporal models, as introduced in "Adaptive Markovian Spatiotemporal Transfer Learning in Multivariate Bayesian Modeling" (Presicce and Banerjee, 2026+) \doi{10.48550/arXiv.2602.08544}. This methodology enables efficient Bayesian multivariate spatiotemporal modeling, utilizing dynamic predictive stacking to improve inference across multivariate time series of spatial datasets. The core functions leverage 'C++' for high-performance computation, making the framework well-suited for large-scale spatiotemporal data analysis in parallel computing environments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lucapresicce.github.io/spFFBS/}
}

}
\author{
\strong{Maintainer}: Luca Presicce \email{l.presicce@campus.unimib.it} (\href{https://orcid.org/0009-0005-7062-3523}{ORCID})

}
\keyword{internal}
